+++ CHANGELOG.md
@@ -54,3 +54,6 @@
 * Define `File` for animations, audio, documents, photos, videos or voices
 * Extract `Size` datatype into separated module within `Content` submodule
 * Define `Location` and `Info` datatypes within `Content` submodule
+
+# 0.1.8
+* Add information about `Group` or `Supergroup` (id and title) into `Moving` datatype